Dissociate widget class and procedure's namespace
BWidget codebase assumes that megawidget's code exists in a namespace named
exactly like the class.  This forbids having megawidget code in lower-cased
namespaces, as widget classes must be upper-case in order to avoid confusing
the option database.

This change adds a new option -namespace to Widget::define, that allows
specifying the namespace where widget code lives, when it is different from the
class name.  The widget dispatcher and the use procedure are now created in
that namespace as well.  Compatibility is maintained with BWidget's
lazy-loading (widgets are not defined when invoking their use procedure).